Add extra checks for gcc-Debug builds
+ Adding '-Wconversion -Wdouble-promotion -Wshadow -Wformat=2' and fix
  issues related to these warnings.
+ Changes recommended by Jason Turner (C++ class 2019-Oct).

+++ src/c4/Timer.hh
@@ -327,7 +327,7 @@ double Timer::system_cpu() const {
 #if defined(WIN32)
   return 0.0; // difftime( tms_end, tms_begin );
 #else
-  return (tms_end.tms_stime - tms_begin.tms_stime) /
+  return static_cast<double>(tms_end.tms_stime - tms_begin.tms_stime) /
          static_cast<double>(posix_clock_ticks_per_second);
 #endif
 }
@@ -341,7 +341,7 @@ double Timer::user_cpu() const {
   duration<double> diff = tms_end - tms_begin;
   return duration_cast<nanoseconds>(diff).count() / 1.0e9;
 #else
-  return (tms_end.tms_utime - tms_begin.tms_utime) /
+  return static_cast<double>(tms_end.tms_utime - tms_begin.tms_utime) /
          static_cast<double>(posix_clock_ticks_per_second);
 #endif
 }
