+++ src/quadrature/gauleg.hh
@@ -18,7 +18,7 @@
 
 namespace rtt_quadrature {
 //---------------------------------------------------------------------------//
-/*! 
+/*!
  * \brief Gauss-Legendre quadrature
  *
  * Calculate abscissae and weights for Gauss-Legendre quadrature:
@@ -43,14 +43,14 @@ namespace rtt_quadrature {
  * This routine scales the range of integration from \f$(x_1,x_2)\f$ to (-1,1)
  * and provides the abscissas \f$ x_j\f$ and weights \f$ w_j \f$ for the
  * Gaussian formula provided above.
- * 
+ *
  * \tparam FieldVector A random access container on a field type.
  *
  * \param x1 Start of integration interval
  * \param x2 End of integration interval
  * \param x On return, contains abscissae \f$x_j\f$ for quadrature.
  * \param w On return, contains weights \f$w_j\f$ for quadrature.
- * \param n Number of points in quadrature. 
+ * \param n Number of points in quadrature.
  */
 template <typename FieldVector>
 void gauleg(
@@ -86,7 +86,8 @@ void gauleg(
   // Loop over the desired roots.
   for (size_t iroot = 0; iroot < numHrGaussPoints; ++iroot) {
     // Approximate the i-th root.
-    Field z(cos(PI * (iroot + 0.75) / (n + 0.5)));
+    Field z(cos(PI * (static_cast<double>(iroot) + 0.75) /
+                (static_cast<double>(n) + 0.5)));
 
     // Temporary storage;
     Field z1, pp;
