+++ src/special_functions/ExpInt.cc
@@ -148,8 +148,8 @@ double Ei(double const x) {
     fact = 1.0;
     size_t k(1);
     for (; k <= MAXIT; ++k) {
-      fact *= x / k;
-      term = fact / k;
+      fact *= x / static_cast<double>(k);
+      term = fact / static_cast<double>(k);
       sum += term;
       if (term < EPS * sum)
         break;
@@ -162,7 +162,7 @@ double Ei(double const x) {
     term = 1.0; //starts with second term
     for (size_t k = 1; k <= MAXIT; ++k) {
       prev = term;
-      term *= k / x;
+      term *= static_cast<double>(k) / x;
       if (term < EPS)
         break; //term approximates relative error
